refactor(app): move freezed files
